Update the sensor framework to report stats back to the HNP if requested by including the data in heartbeats.
This commit was SVN r27748.

+++ orte/mca/sensor/ft_tester/sensor_ft_tester.c
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2009-2011 Cisco Systems, Inc.  All rights reserved. 
- * Copyright (c) 2011      Los Alamos National Security, LLC.
+ * Copyright (c) 2011-2012 Los Alamos National Security, LLC.
  *                         All rights reserved.
  *
  * $COPYRIGHT$
@@ -28,13 +28,10 @@
 
 #include "opal_stdint.h"
 #include "opal/util/output.h"
-#include "opal/mca/event/event.h"
 
 #include "orte/util/error_strings.h"
 #include "orte/util/name_fns.h"
 #include "orte/mca/errmgr/errmgr.h"
-#include "orte/mca/odls/odls.h"
-#include "orte/mca/odls/base/odls_private.h"
 #include "orte/runtime/orte_globals.h"
 
 #include "orte/mca/sensor/base/base.h"
@@ -42,91 +39,31 @@
 #include "sensor_ft_tester.h"
 
 /* declare the API functions */
-static int init(void);
-static void finalize(void);
-static void start(orte_jobid_t job);
-static void stop(orte_jobid_t job);
+static void sample(void);
 
 /* instantiate the module */
 orte_sensor_base_module_t orte_sensor_ft_tester_module = {
-    init,
-    finalize,
-    start,
-    stop
+    NULL,
+    NULL,
+    NULL,
+    NULL,
+    sample,
+    NULL
 };
 
-/* declare the local functions */
-static void sample(int fd, short event, void *arg);
-
-/* local globals */
-static opal_event_t *sample_ev = NULL;
-static struct timeval sample_time;
-
-static int init(void)
-{
-    if (0 == mca_sensor_ft_tester_component.fail_rate) {
-        /* not monitoring */
-        return ORTE_ERROR;
-    }
-
-    return ORTE_SUCCESS;
-}
-
-static void finalize(void)
-{
-    if (NULL != sample_ev) {
-        opal_event_del(sample_ev);
-        free(sample_ev);
-        sample_ev = NULL;
-    }
-    
-    return;
-}
-
-/*
- * Start killing local processes
- */
-static void start(orte_jobid_t jobid)
-{
-    if (NULL == sample_ev) {
-        /* startup a timer to wake us up periodically */
-        sample_ev =  (opal_event_t *) malloc(sizeof(opal_event_t));
-        opal_event_evtimer_set(orte_event_base, sample_ev, sample, sample_ev);
-        sample_time.tv_sec = mca_sensor_ft_tester_component.fail_rate;
-        sample_time.tv_usec = 0;
-        opal_event_evtimer_add(sample_ev, &sample_time);
-    }
-    return;
-}
-
-
-static void stop(orte_jobid_t jobid)
-{
-    if (NULL != sample_ev) {
-        opal_event_del(sample_ev);
-        free(sample_ev);
-        sample_ev = NULL;
-    }
-    return;
-}
-
-static void sample(int fd, short event, void *arg)
+static void sample(void)
 {
     float prob;
     orte_proc_t *child;
     int i;
 
-    /* if we are not sampling any more, then just return */
-    if (NULL == sample_ev) {
-        return;
-    }
-    
     OPAL_OUTPUT_VERBOSE((1, orte_sensor_base.output,
                          "%s sample:ft_tester considering killing something",
                          ORTE_NAME_PRINT(ORTE_PROC_MY_NAME)));
 
     /* are we including ourselves? */
-    if (ORTE_PROC_IS_DAEMON && 0 < mca_sensor_ft_tester_component.daemon_fail_prob) {
+    if ((ORTE_PROC_IS_DAEMON || ORTE_PROC_IS_CMSLAVE) &&
+        0 < mca_sensor_ft_tester_component.daemon_fail_prob) {
         OPAL_OUTPUT_VERBOSE((1, orte_sensor_base.output,
                              "%s sample:ft_tester considering killing me!",
                              ORTE_NAME_PRINT(ORTE_PROC_MY_NAME)));
@@ -142,44 +79,41 @@ static void sample(int fd, short event, void *arg)
         }
     }
 
-    /* see if we should kill a child */
-    for (i=0; i < orte_local_children->size; i++) {
-        if (NULL == (child = (orte_proc_t*)opal_pointer_array_get_item(orte_local_children, i))) {
-            continue;
-        }
-        if (!child->alive || 0 == child->pid ||
-            ORTE_PROC_STATE_UNTERMINATED < child->state) {
+    if (0 < mca_sensor_ft_tester_component.fail_prob) {
+        /* see if we should kill a child */
+        for (i=0; i < orte_local_children->size; i++) {
+            if (NULL == (child = (orte_proc_t*)opal_pointer_array_get_item(orte_local_children, i))) {
+                continue;
+            }
+            if (!child->alive || 0 == child->pid ||
+                ORTE_PROC_STATE_UNTERMINATED < child->state) {
+                OPAL_OUTPUT_VERBOSE((1, orte_sensor_base.output,
+                                     "%s sample:ft_tester ignoring child: %s alive %s pid %lu state %s",
+                                     ORTE_NAME_PRINT(ORTE_PROC_MY_NAME),
+                                     ORTE_NAME_PRINT(&child->name),
+                                     child->alive ? "TRUE" : "FALSE",
+                                     (unsigned long)child->pid, orte_proc_state_to_str(child->state)));
+                continue;
+            }
+            /* roll the dice */
+            prob = (double)random() / (double)INT32_MAX;
             OPAL_OUTPUT_VERBOSE((1, orte_sensor_base.output,
-                                 "%s sample:ft_tester ignoring child: %s alive %s pid %lu state %s",
+                                 "%s sample:ft_tester child: %s dice: %f prob %f",
                                  ORTE_NAME_PRINT(ORTE_PROC_MY_NAME),
                                  ORTE_NAME_PRINT(&child->name),
-                                 child->alive ? "TRUE" : "FALSE",
-                                 (unsigned long)child->pid, orte_proc_state_to_str(child->state)));
-            continue;
-        }
-        /* roll the dice */
-        prob = (double)random() / (double)INT32_MAX;
-        OPAL_OUTPUT_VERBOSE((1, orte_sensor_base.output,
-                             "%s sample:ft_tester child: %s dice: %f prob %f",
-                             ORTE_NAME_PRINT(ORTE_PROC_MY_NAME),
-                             ORTE_NAME_PRINT(&child->name),
-                             prob, mca_sensor_ft_tester_component.fail_prob));
-        if (prob < mca_sensor_ft_tester_component.fail_prob) {
-            /* you shall die... */
-            OPAL_OUTPUT_VERBOSE((1, orte_sensor_base.output,
-                                 "%s sample:ft_tester killing %s",
-                                 ORTE_NAME_PRINT(ORTE_PROC_MY_NAME),
-                                 ORTE_NAME_PRINT(&child->name)));
-            kill(child->pid, SIGTERM);
-            /* are we allowing multiple deaths */
-            if (!mca_sensor_ft_tester_component.multi_fail) {
-                break;
+                                 prob, mca_sensor_ft_tester_component.fail_prob));
+            if (prob < mca_sensor_ft_tester_component.fail_prob) {
+                /* you shall die... */
+                OPAL_OUTPUT_VERBOSE((1, orte_sensor_base.output,
+                                     "%s sample:ft_tester killing %s",
+                                     ORTE_NAME_PRINT(ORTE_PROC_MY_NAME),
+                                     ORTE_NAME_PRINT(&child->name)));
+                kill(child->pid, SIGTERM);
+                /* are we allowing multiple deaths */
+                if (!mca_sensor_ft_tester_component.multi_fail) {
+                    break;
+                }
             }
         }
-    }
-
-    /* restart the timer */
-    if (NULL != sample_ev) {
-        opal_event_evtimer_add(sample_ev, &sample_time);
     }
 }
